refs #155 fix bunch size parsing

+++ app/src/main/java/net/eneiluj/nextcloud/phonetrack/android/fragment/PreferencesFragment.java
@@ -235,7 +235,7 @@ public class PreferencesFragment extends PreferenceFragmentCompat implements Pre
 
         final EditTextPreference groupSyncPref = (EditTextPreference) findPreference(getString(R.string.pref_key_group_sync));
         String groupSyncValStr = sp.getString(getString(R.string.pref_key_group_sync), "0");
-        int groupSyncVal = Integer.valueOf(groupSyncValStr);
+        long groupSyncVal = Long.valueOf(groupSyncValStr);
         groupSyncPref.setSummary(String.valueOf(groupSyncVal));
         groupSyncPref.setDialogMessage(getString(R.string.settings_group_sync_long));
         groupSyncPref.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
@@ -249,7 +249,7 @@ public class PreferencesFragment extends PreferenceFragmentCompat implements Pre
                     return false;
                 }
                 else {
-                    int valInt = Integer.valueOf(newValueString);
+                    long valInt = Long.valueOf(newValueString);
                     //groupSyncPref.setText(String.valueOf(valInt));
                     /*SharedPreferences.Editor editor = sp.edit();
                     editor.putString(getString(R.string.pref_key_group_sync), String.valueOf(valInt));
